pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.61.0",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "streamlit-lottie>=0.0.5",
    "streamlit>=1.41.1",
    "trafilatura>=2.0.0",
    "google-generativeai>=0.8.4",
]
